Split the variant test into six translation units to fit into the CI memory limits; add a simple constexpr test
